Pan and tilt values updated via state controller
When controlling the head with a different controller, the head moved with a very high speed, which can be harmful to the hardware.
It was assumed that the head was only moved via teleop, thus the position was not updated when the head was moved with a different controller.
This was fixed by reading the actual values for pan and tilt from the state controller.