Only warn the user on outdated repeated messages
dornhege's commit recently added a safety mechanism to ensure
a properly working dead-man switch with the PS2 controller.

However, in some setups this problem appears during normal operation
but does not persist for long. In order to avoid spamming the log
console, we warn the user only if the received messages are additionally
out of date.

The overall safety mechanism still triggers, this only affects the log.